+++ all-is-cubes-ui/src/vui/widgets/tooltip.rs
@@ -312,13 +312,20 @@ impl WidgetController for TooltipController {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use all_is_cubes::transaction::{self, Transaction as _};
+    use all_is_cubes::universe::UniverseTransaction;
     use all_is_cubes::util::yield_progress_for_testing;
 
     #[tokio::test]
     async fn tooltip_timeout_and_dirty_text() {
         // TODO: reduce boilerplate
+
         let mut universe = Universe::new();
-        let hud_blocks = &HudBlocks::new(&mut universe, yield_progress_for_testing()).await;
+        let mut install_txn = UniverseTransaction::default();
+        let hud_blocks = &HudBlocks::new(&mut install_txn, yield_progress_for_testing()).await;
+        install_txn
+            .execute(&mut universe, &mut transaction::no_outputs)
+            .unwrap();
 
         // Initial state: no update.
         let mut t = TooltipState::default();